support SA partitioning
